pyproject.toml
[project]
name = "txt2svg"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "diffusers",
    "transformers",
    "accelerate",
    "sentencepiece",
    "protobuf",
    "bitsandbytes",
    "scikit-image",
    "opencv-python",
    "numpy",
    "Pillow",
    "networkx",
    "scipy",
    "vpype",
    "openai",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
include = ["modules*"]
